Fixed compiler warning/error
git-svn-id: https://qlc.svn.sourceforge.net/svnroot/qlc/trunk@3008 0db28fee-5c31-0410-96ea-e69cce70e049

+++ engine/src/chaserrunner.cpp
@@ -198,7 +198,7 @@ void ChaserRunner::previous()
 
 void ChaserRunner::tap()
 {
-    if (m_roundTime->elapsed() >= (currentDuration() / 4))
+    if (uint(m_roundTime->elapsed()) >= (currentDuration() / 4))
         next();
 }
